sort by date

+++ src/utils.ts
@@ -28,7 +28,7 @@ export function formatDate(isoDate: string | undefined): string {
 	if(!isoDate) return ""
 
 	const date = new Date(isoDate);
-	const options: Intl.DateTimeFormatOptions = { year: "numeric", month: "long", day: "numeric" };
+	const options: Intl.DateTimeFormatOptions = { year: "numeric", month: "numeric", day: "numeric" };
 	return date.toLocaleDateString(undefined, options);
 }
